# apps/api/src/app.ts
import { URL } from "node:url";
import express from "express";
import cookieParser from "cookie-parser";
import cors from "cors";
import helmet from "helmet";
import dotenv from "dotenv";
dotenv.config();
import { validateEnv, type Env } from "./config/env";
import { getDatabaseClient } from "@handoverkey/database";

let validatedEnv: Env | undefined;
if (process.env.NODE_ENV !== "test") {
  validatedEnv = validateEnv();
}

const app = express();
app.set("trust proxy", 1);

import {
  rateLimiter,
  validateContentType,
  sanitizeInput,
} from "./middleware/security";
import { errorHandler, notFoundHandler } from "./middleware/error-handler";
import { requestIdMiddleware } from "./middleware/request-id";
import { loggingMiddleware } from "./middleware/logging";
import { metricsMiddleware } from "./middleware/metrics";
import { logger } from "./config/logger";
import { getMetrics, getMetricsContentType } from "./config/metrics";
import authRoutes from "./routes/auth-routes";
import vaultRoutes, { publicVaultRouter } from "./routes/vault-routes";
import activityRoutes, { publicActivityRouter } from "./routes/activity-routes";
import inactivityRoutes from "./routes/inactivity-routes";
import sessionRoutes from "./routes/session-routes";
import successorRoutes, { verifyRouter } from "./routes/successor-routes";
import handoverRoutes, { publicHandoverRouter } from "./routes/handover-routes";
import adminRoutes from "./routes/admin-routes";
import contactRoutes from "./routes/contact-routes";
import waitlistRoutes from "./routes/waitlist-routes";
import billingRoutes from "./routes/billing-routes";
import { BillingController } from "./controllers/billing-controller";
import { JobProcessor, JobScheduler } from "./jobs";
import { closeAllQueues, getQueueHealth } from "./config/queue";
import { SessionService } from "./services/session-service";
import { initializeRedis, closeRedis, checkRedisHealth } from "./config/redis";
import { JobManager } from "./services/job-manager";
import { realtimeService } from "./services/realtime-service";
import { authenticateJWT, requireAuth } from "./middleware/auth";

const dbClient = getDatabaseClient();
const jobManager = JobManager.getInstance();
export let appInit: Promise<void> = Promise.resolve();

if (process.env.NODE_ENV !== "test" && validatedEnv) {
  appInit = dbClient
    .initialize({
      host: validatedEnv.DB_HOST,
      port: validatedEnv.DB_PORT,
      database: validatedEnv.DB_NAME,
      user: validatedEnv.DB_USER,
      password: validatedEnv.DB_PASSWORD,
      min: 2,
      max: 10,
    })
    .then(async () => {
      // Initialize SessionService with database client
      SessionService.initialize(dbClient);
      logger.info("SessionService initialized");

      // Initialize Redis
      await initializeRedis();
      logger.info("Redis initialized");

      // Initialize job processors and schedule recurring jobs
      await JobProcessor.initialize();
      await JobScheduler.scheduleInactivityCheck();
      await JobScheduler.scheduleSessionCleanup();
      await JobScheduler.scheduleArchiveLogs();
      jobManager.start();
      logger.info("Job processors initialized and recurring jobs scheduled");
    })
    .catch((error) => {
      logger.fatal({ err: error }, "Failed to initialize application");
      process.exit(1);
    });
}

// Request ID middleware (must be first)
app.use(requestIdMiddleware);

app.use(
  helmet({
    contentSecurityPolicy: {
      directives: {
        defaultSrc: ["'self'"],
        scriptSrc: ["'self'"],
        styleSrc: ["'self'", "'unsafe-inline'"],
        imgSrc: ["'self'", "data:"],
        connectSrc: ["'self'"],
        fontSrc: ["'self'"],
        objectSrc: ["'none'"],
        frameAncestors: ["'none'"],
        formAction: ["'self'"],
        baseUri: ["'self'"],
      },
    },
    crossOriginEmbedderPolicy: false,
  }) as unknown as express.RequestHandler,
);

// Build the CORS allowlist. For each configured origin we also allow the
// www / non-www counterpart (apex domains only) so that
// FRONTEND_URL=https://handoverkey.app automatically permits
// https://www.handoverkey.app and vice-versa.
export function buildAllowedOrigins(raw: string): Set<string> {
  const origins = new Set<string>();

  for (const entry of raw.split(",")) {
    const trimmed = entry.trim();
    if (!trimmed) continue;

    let parsed: URL;
    try {
      parsed = new URL(trimmed);
    } catch {
      logger.warn({ origin: trimmed }, "Ignoring malformed CORS origin");
      continue;
    }

    // url.origin normalises scheme + host + port (strips paths, trailing slashes)
    origins.add(parsed.origin);

    const parts = parsed.hostname.split(".");
    if (parsed.hostname.startsWith("www.") && parts.length === 3) {
      // www.example.com -> also allow example.com
      const bare = new URL(parsed.origin);
      bare.hostname = parsed.hostname.slice(4);
      origins.add(bare.origin);
    } else if (parts.length === 2) {
      // example.com -> also allow www.example.com
      const www = new URL(parsed.origin);
      www.hostname = `www.${parsed.hostname}`;
      origins.add(www.origin);
    }
  }

  return origins;
}

const ALLOWED_ORIGINS = buildAllowedOrigins(
  process.env.CORS_ORIGINS ||
    process.env.FRONTEND_URL ||
    "http://localhost:5173",
);

app.use(
  cors({
    origin: (origin, callback) => {
      if (!origin || ALLOWED_ORIGINS.has(origin)) {
        callback(null, true);
      } else {
        callback(new Error(`Origin ${origin} not allowed by CORS`));
      }
    },
    credentials: true,
    methods: ["GET", "POST", "PUT", "DELETE", "OPTIONS", "PATCH"],
    allowedHeaders: [
      "Content-Type",
      "Authorization",
      "X-Requested-With",
      "Accept",
      "Origin",
      "DNT",
      "Cache-Control",
      "X-Request-ID",
    ],
  }),
);

// Logging middleware (must be early)
app.use(loggingMiddleware);

// Metrics middleware (must be early)
app.use(metricsMiddleware);

// Security middleware (rate limiter only)
app.use(rateLimiter as unknown as express.RequestHandler);

// Stripe webhook needs raw body — must be registered before JSON body parser
app.post(
  "/api/v1/billing/webhook",
  express.raw({ type: "application/json" }) as express.RequestHandler,
  BillingController.handleWebhook as unknown as express.RequestHandler,
);

// Body parsing middleware (capped at 5MB per security policy)
app.use(express.json({ limit: "5mb" }) as express.RequestHandler);
app.use(
  express.urlencoded({
    extended: true,
    limit: "5mb",
  }) as express.RequestHandler,
);
app.use(cookieParser() as unknown as express.RequestHandler);

// Validation and sanitization middleware
app.use(validateContentType);
app.use(sanitizeInput);

const healthCheckHandler: express.RequestHandler = async (_req, res) => {
  try {
    const [dbHealthy, redisHealthy] = await Promise.all([
      dbClient.healthCheck(),
      checkRedisHealth(),
    ]);

    const queueNames = [
      "inactivity",
      "notifications",
      "handover",
      "maintenance",
    ];
    const queueHealthResults = await Promise.all(
      queueNames.map((name) => getQueueHealth(name)),
    );
    const jobQueueHealthy = queueHealthResults.every(
      (result) => result.healthy,
    );
    const managerHealth = await jobManager.getHealthStatus();

    const allHealthy =
      dbHealthy && redisHealthy && jobQueueHealthy && managerHealth.isHealthy;

    res.status(allHealthy ? 200 : 503).json({
      status: allHealthy ? "ok" : "degraded",
      timestamp: new Date().toISOString(),
      version: process.env.npm_package_version || "0.1.0",
      checks: {
        database: dbHealthy ? "ok" : "failed",
        redis: redisHealthy ? "ok" : "failed",
        jobQueue: jobQueueHealthy ? "ok" : "degraded",
        inactivityMonitor: managerHealth.jobs.inactivityMonitor.isHealthy
          ? "ok"
          : "degraded",
      },
      queueStats: queueNames.reduce<Record<string, unknown>>(
        (acc, name, index) => {
          acc[name] = queueHealthResults[index];
          return acc;
        },
        {},
      ),
      backgroundJobs: managerHealth.jobs,
      realtime: realtimeService.getStats(),
    });
  } catch {
    res.status(503).json({
      status: "error",
      timestamp: new Date().toISOString(),
      version: process.env.npm_package_version || "0.1.0",
      error: "Health check failed",
    });
  }
};

// Health check endpoints
app.get("/health", healthCheckHandler);
app.get("/api/health", healthCheckHandler);

// Prometheus metrics endpoint (protected -- requires authentication)
app.get("/metrics", authenticateJWT, requireAuth, async (req, res) => {
  try {
    res.set("Content-Type", getMetricsContentType());
    const metrics = await getMetrics();
    res.send(metrics);
  } catch (error) {
    logger.error({ err: error }, "Failed to generate metrics");
    res.status(500).send("Failed to generate metrics");
  }
});

// API routes
app.use("/api/v1/auth", authRoutes);
app.use("/api/v1/vault", publicVaultRouter);
app.use("/api/v1/vault", vaultRoutes);
app.use("/api/v1/activity", publicActivityRouter);
app.use("/api/v1/activity", activityRoutes);
app.use("/api/v1/inactivity", inactivityRoutes);
app.use("/api/v1/sessions", sessionRoutes);
app.use("/api/v1/successors", verifyRouter); // Public verify route (no auth)
app.use("/api/v1/successors", successorRoutes); // Protected routes
app.use("/api/v1/handover", publicHandoverRouter); // Public respond route
app.use("/api/v1/handover", handoverRoutes); // Protected routes
app.use("/api/v1/admin", adminRoutes);
app.use("/api/v1/contact", contactRoutes);
app.use("/api/v1/waitlist", waitlistRoutes);
app.use("/api/v1/billing", billingRoutes);

// 404 handler - must be after all routes
app.use(notFoundHandler);

// Global error handler - must be last
app.use(errorHandler);

/**
 * Shuts down background services (jobs, Redis, database).
 * Called from index.ts after the HTTP server is closed.
 */
export async function shutdownServices(): Promise<void> {
  jobManager.stop();

  const shutdownErrors: string[] = [];

  const shutdownSteps: Array<{
    name: string;
    run: () => Promise<void>;
  }> = [
    { name: "JobProcessor", run: () => JobProcessor.close() },
    { name: "Queues", run: () => closeAllQueues() },
    { name: "Redis", run: () => closeRedis() },
    { name: "Database", run: () => dbClient.close() },
  ];

  for (const step of shutdownSteps) {
    try {
      await step.run();
    } catch (error) {
      const reason = error instanceof Error ? error.message : String(error);
      shutdownErrors.push(`${step.name}: ${reason}`);
    }
  }

  if (shutdownErrors.length > 0) {
    throw new Error(
      `One or more services failed to shutdown cleanly: ${shutdownErrors.join("; ")}`,
    );
  }
}

export default app;