# docs/openapi.yaml
openapi: 3.1.0
info:
  title: HandoverKey API
  description: >
    REST API for the HandoverKey digital legacy platform. All protected
    endpoints require a valid JWT access token sent as a Bearer header or
    httpOnly cookie.
  version: 1.0.0
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  contact:
    name: HandoverKey
    url: https://handoverkey.app
servers:
  - url: /api/v1
    description: Default (relative)
  - url: http://localhost:3001/api/v1
    description: Local development
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    cookieAuth:
      type: apiKey
      in: cookie
      name: accessToken
...
security:
  - bearerAuth: []
  - cookieAuth: []
paths:
...
  /successors:
    get:
      tags: [Successors]
      summary: List successors
      responses:
        "200":
          description: Array of successors
    post:
      tags: [Successors]
      summary: Add a successor
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, email]
              properties:
                name:
                  type: string
                email:
                  type: string
                  format: email
                handoverDelayDays:
                  type: integer
      responses:
        "201":
          description: Successor added
...
  /activity:
    get:
      tags: [Activity]
      summary: Get activity logs
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        "200":
          description: Paginated activity logs
...
  /contact:
    post:
      tags: [Contact]
      summary: Submit contact form
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, email, message]
              properties:
                name:
                  type: string
                email:
                  type: string
                  format: email
                message:
                  type: string
      responses:
        "200":
          description: Message sent